backend/requirements.txt
beautifulsoup4==4.12.2
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
colorama==0.4.6
deprecation==2.1.0
Flask==2.2.3
functions-framework==3.3.0
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-bigquery==1.26.1
google-cloud-core==1.7.3
google-cloud-storage==1.30.0
google-crc32c==1.5.0
google-resumable-media==0.6.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.2
packaging==23.0
pandas==2.0.0
protobuf==3.20.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
rsa==4.9
scikit-learn==1.0.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==1.26.15
watchdog==3.0.0
Werkzeug==2.2.3
zipp==3.15.0
